+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- Added metrics exporter
+  ([#678](https://github.com/census-instrumentation/opencensus-python/pull/678)
 
 ## 0.2.1
 Released 2019-06-13